+++ hutool-core/src/test/java/com/xiaoleilu/hutool/core/io/WatchMonitorTest.java
@@ -1,0 +1,16 @@
+package com.xiaoleilu.hutool.core.io;
+
+import org.junit.Test;
+
+/**
+ * 文件监听单元测试
+ * @author Looly
+ *
+ */
+public class WatchMonitorTest {
+	
+	@Test
+	public void watchMonitorTest(){
+		
+	}
+}